refactor(vault): extract note-ops module, reduce vault.ts to pure barrel
Move the entire mutation layer (writeNote/updateNote/appendToNote/
prependToNote/deleteNote/moveNote, their option/result interfaces, and the
private frontmatter/trash/cross-device-move helpers) into
src/lib/note-ops.ts, which depends on vault-fs, note-reads, and
link-rewriter. vault.ts now contains only re-export statements, so the
./vault.js public API is unchanged while holding zero logic.
